Update 030421 1
First crack at native netplay
This is only half implemented currently, and it uses the available command line options right now to kick off netplay sessions.  This might not be long term the right way to go, the other option would be to generate a temporary retroarch.cfg and appendconfig.  The reason being that a lot more netplay options are avialable with that method.  Going to try out this method first and see how it goes in testing